requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.9
whitenoise==5.0.1
